Fix -Wsign-compare warning

+++ lib/modular_arithmetic/montgomery_arithmetic/include/hurchalla/montgomery_arithmetic/detail/experimental/montgomery_pow_kary.h
@@ -145,7 +145,7 @@ typename MF::MontgomeryValue montgomery_pow_kary(const MF& mf, typename MF::Mont
     int leading_zeros = count_leading_zeros(n);
     int numbits = ut_numeric_limits<decltype(n)>::digits - leading_zeros;
     // because we returned above if (n <= MASK), we can assert the following:
-    HPBC_ASSERT(numbits > P);
+    HPBC_ASSERT(numbits > static_cast<int>(P));
 
     int shift = numbits - static_cast<int>(P);
     U tmp = n >> shift;
@@ -175,7 +175,7 @@ typename MF::MontgomeryValue montgomery_pow_kary(const MF& mf, typename MF::Mont
 
     if (shift == 0)
         return result;
-    HPBC_ASSERT(0 < shift && shift < P);
+    HPBC_ASSERT(0 < shift && shift < static_cast<int>(P));
 
     for (int i=0; i<shift; ++i)
         result = mf.square(result);
@@ -265,7 +265,7 @@ array_montgomery_pow_kary(const std::array<MF, ARRAY_SIZE>& mf,
     int leading_zeros = count_leading_zeros(n_max);
     int numbits = ut_numeric_limits<decltype(n_max)>::digits - leading_zeros;
     // because we returned above if (n_max <= MASK), we can assert the following:
-    HPBC_ASSERT(numbits > P);
+    HPBC_ASSERT(numbits > static_cast<int>(P));
 
     int shift = numbits - static_cast<int>(P);
     std::array<V, ARRAY_SIZE> result;
@@ -295,7 +295,7 @@ array_montgomery_pow_kary(const std::array<MF, ARRAY_SIZE>& mf,
 
     if (shift == 0)
         return result;
-    HPBC_ASSERT(0 < shift && shift < P);
+    HPBC_ASSERT(0 < shift && shift < static_cast<int>(P));
 
     for (int i=0; i<shift; ++i) {
         HURCHALLA_REQUEST_UNROLL_LOOP for (size_t j=0; j<ARRAY_SIZE; ++j)
